inference-web-app/requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
cachetools==5.3.1
category-encoders==2.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
patsy==0.5.3
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.3.5
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
statsmodels==0.14.0
streamlit==1.22.0
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.20.0
xgboost==1.7.5
zipp==3.15.0